--- apps/pxweb2/src/app/components/NavigationBar/NavigationBar.module.scss
@@ -1,13 +1,0 @@
-@use '$ui/style-dictionary/dist/scss/fixed-variables.scss' as fixed;
-
-.navigationBar {
-  height: 80px;
-  margin: 0;
-  padding: 16px;
-  padding-bottom: 32px;
-  width: 100vw;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  background-color: var(--px-color-background-default);
-}
