+++ apps/pxweb2/src/app/components/NavigationMenu/NavigationBar/NavigationBar.module.scss
@@ -1,0 +1,13 @@
+@use '$ui/style-dictionary/dist/scss/fixed-variables.scss' as fixed;
+
+.navigationBar {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 80px;
+  padding: 0px fixed.$spacing-2;
+  gap: fixed.$spacing-2;
+
+  background-color: var(--px-color-surface-subtle);
+}
