+++ apps/pxweb2/src/app/components/NavigationMenu/NavigationItem/NavigationItem.module.scss
@@ -1,24 +1,10 @@
 @use '$ui/style-dictionary/dist/scss/fixed-variables.scss' as fixed;
 
-/** Styling */
-.navigationRail {
-  display: flex;
-  width: 120px;
-  height: 100%;
-  padding: 32px 0px 56px 0px;
-  flex-direction: column;
-  align-items: center;
-  gap: 24px;
-
-  background-color: var(--px-color-surface-default);
-}
-
 .item {
   display: flex;
   flex-direction: column;
   align-items: center;
-  width: 100px;
-  padding: 0 0 fixed.$spacing-1 0;
+  
   gap: fixed.$spacing-1;
   border-radius: var(--px-border-radius-small);
   cursor: pointer;
@@ -31,7 +17,7 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    width: fixed.$spacing-14;
+
     padding-top: fixed.$spacing-1;
     padding-bottom: fixed.$spacing-1;
     border-radius: 99px;
@@ -54,5 +40,25 @@
     > .icon {
       background-color: var(--px-color-surface-action-subtle-active);
     }
+  }
+}
+
+.navBarItem {
+  width: 100%;
+  padding: fixed.$spacing-2 0px;
+  justify-content: center;
+  flex: 1 0 0;
+
+  > .icon {
+    width: fixed.$spacing-16;
+  }
+}
+
+.navRailItem {
+  width: 100px;
+  padding: 0px 0px fixed.$spacing-1 0px;
+
+  > .icon {
+    width: fixed.$spacing-14;
   }
 }
