+++ apps/pxweb2/src/app/components/NavigationMenu/NavigationItem/NavigationItemType.ts
@@ -1,0 +1,7 @@
+export type NavigationItem =
+  | 'none'
+  | 'filter'
+  | 'view'
+  | 'edit'
+  | 'save'
+  | 'help';
